Refactor: 누락된 components 디렉토리를 생성한다 (#9)
