+++ apps/landing/src/components/common/DefaultLayout/DefaultLayout.component.tsx
