+++ apps/landing/src/components/common/DefaultLayout/DefaultLayout.styled.ts
