+++ apps/landing/src/components/common/Header/Header.component.tsx
@@ -1,5 +1,5 @@
 import * as Styled from './Header.styled';
-import { Heading, Navigation } from '@/common';
+import { Heading, Navigation } from '@/components/common';
 
 const Header = () => (
   <Styled.Header>
